requirements.txt
pandas
numpy
scipy
ydata-profiling
streamlit-ydata-profiling
matplotlib
scikit-learn
